+++ packages/moonstone/Scroller/Scrollbar.js
@@ -211,7 +211,9 @@ class ScrollbarBase extends PureComponent {
 			currentPos = vertical ? bounds.scrollTop : bounds.scrollLeft,
 			maxPos = vertical ? bounds.maxTop : bounds.maxLeft,
 			shouldDisablePrevButton = currentPos <= 0,
-			shouldDisableNextButton = currentPos >= maxPos,
+			/* If a scroll size or a client size is not integer,
+			   browsers's max scroll position could be smaller than maxPos by 1 pixel.*/
+			shouldDisableNextButton = maxPos - currentPos <= 1,
 			spotItem = Spotlight.getCurrent();
 
 		this.setState((prevState) => {