+++ packages/moonstone/VirtualList/VirtualListBaseNative.js
@@ -392,10 +392,9 @@ class VirtualListCoreNative extends Component {
 
 	getItemPosition = (index, stickTo = 'start') => {
 		const
-			{itemSize} = this.props,
 			{primary} = this,
 			position = this.getGridPosition(index),
-			offset = ((itemSize instanceof Object) || stickTo === 'start') ? 0 : primary.clientSize - primary.itemSize;
+			offset = (stickTo === 'start') ? 0 : primary.clientSize - primary.itemSize;
 
 		position.primaryPosition -= offset;
 
@@ -792,41 +791,28 @@ class VirtualListCoreNative extends Component {
 		this.setRestrict(isSelfOnly);
 	}
 
-	getIndicesForPageScroll = (direction, currentIndex) => {
+	getIndexForPageScroll = (direction, currentIndex) => {
 		const
 			{context, dimensionToExtent, isPrimaryDirectionVertical, primary} = this,
-			{dataSize} = this.props,
-			indices = {};
+			{dataSize, spacing} = this.props;
+		let offsetIndex = Math.floor((primary.clientSize + spacing) / primary.gridSize) * dimensionToExtent;
 
-		let	offsetIndex = Math.floor(primary.clientSize / primary.gridSize) * dimensionToExtent;
 		offsetIndex *= !isPrimaryDirectionVertical && context.rtl ? -1 : 1;
+		offsetIndex *= (direction === 'down' || direction === 'right') ? 1 : -1;
 
-		if (direction === 'down' || direction === 'right') {
-			indices.indexToFocus = clamp(0, dataSize - 1, currentIndex + offsetIndex);
-			indices.indexToScroll = currentIndex + dimensionToExtent;
-			if (context.rtl && !isPrimaryDirectionVertical) {
-				indices.indexToScroll = indices.indexToFocus;
-			}
-		} else {
-			indices.indexToFocus = clamp(0, dataSize - 1, currentIndex - offsetIndex);
-			indices.indexToScroll = indices.indexToFocus;
-			if (context.rtl && !isPrimaryDirectionVertical) {
-				indices.indexToScroll = currentIndex + dimensionToExtent;
-			}
-		}
-
-		return indices;
+		return clamp(0, dataSize - 1, currentIndex + offsetIndex);
 	}
 
 	scrollToNextPage = ({direction, focusedItem}) => {
 		const
+			isRtl = this.context.rtl,
+			isForward = (direction === 'down' || isRtl && direction === 'left' || !isRtl && direction === 'right'),
 			focusedIndex = Number.parseInt(focusedItem.getAttribute(dataIndexAttribute)),
-			{indexToFocus, indexToScroll} = this.getIndicesForPageScroll(direction, focusedIndex);
+			indexToFocus = this.getIndexForPageScroll(direction, focusedIndex);
 
 		if (focusedIndex !== indexToFocus) {
 			focusedItem.blur();
-			this.props.cbScrollTo({index: indexToScroll, animate: false});
-			this.focusByIndex(indexToFocus);
+			this.props.cbScrollTo({index: indexToFocus, stickTo: isForward ? 'end' : 'start', focus: true, animate: false});
 		}
 
 		return true;